requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.7
django-debug-toolbar==5.0.1
django-environ==0.12.0
djangorestframework==3.15.2
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==24.2
psycopg2==2.9.10
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
